backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=7.5.2",
    "fastapi[standard]>=0.115.11",
    "google>=3.0.0",
    "google-genai>=1.5.0",
    "openai>=1.66.3",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "selenium>=4.29.0",
]
